Allow to provide different responses depending on the PHP version
This allows to specify "chances" above 100 percent, that are additional percentages (because we can not use multiple 100 entries).

Background for this is, that we should be able to have a different response for users on a given version depending on their PHP version. In this case Nextcloud 10 is the last version on PHP 5.4 and thus an upgrade to 11 will not be provided. Nevertheless the user should be informed about that problem and that they should upgrade to a newer PHP version.

This commit fixes how that constraints are handled:

* PHP version is now checked in the loop so that we properly skip incompatible ones
* NC version is also checked in the loop for the same reason
* a bug in the config reader is fixed where 9.1 had a fallback to 9.0 and thus serving (accidentially correct) answers
* a integration test is added for the disable autoupdater state

Signed-off-by: Morris Jobke <hey@morrisjobke.de>